+++ src/index.ts
@@ -1,5 +1,4 @@
 import './declarations';
 import '@curveball/links';
 
-export { addSchemasForDir, schemaValidate } from './util';
 export { default as default } from './mw';