+++ src/mw.ts
@@ -2,7 +2,7 @@ import { Middleware } from '@curveball/core';
 import Ajv2019 from 'ajv/dist/2019';
 import addFormats from 'ajv-formats';
 import { findSchemas } from './util';
-import betterAjvErrors from 'better-ajv-errors';
+import * as betterAjvErrors from 'better-ajv-errors';
 import { UnprocessableEntity } from '@curveball/http-errors';
 import { SchemaCollectionController, SchemaController } from './controllers';